Lowered down AWS version
Lowering down the aws version to make it compatible with haystack agent. Haystack agent is using amazon-kinesis-dispatcher which brings aws-java-sdk-core with a version lower than the version in blobs. This makes the version incompatibility problem in haystack-agent with all aws jars on one version and aws-java-sdk-core on another version. We will need to change the version of amazon-kinesis-dispatcher whenever we think of upgrading aws version in blobs